Check Gemfile.lock in the repo
You actually want people to have the same Gemfile.lock because of fixed
gem versions in it. It will provide a smaller amount of versions
depending on each other. It means less dependency pain between different
users or installations of the gem (which we experienced alongside update of
CocoaPods/Xcodeproj to 0.16.0).